Add default system prompt settings
